+++ logstash-output-hyperion/CHANGELOG.md
@@ -1,2 +1,2 @@
 ## 0.1.0
-  - Plugin created with the logstash plugin generator
+- Initial version that can communicate with hyperion pipeline.